wakeup-detection/requirements.txt
numpy==1.19.5
pandas==1.1.5
setuptools==52.0.0
flask==1.1.2
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0